FIXED: Going through hundreds of compile warnings i found that someone still don't know operators precedence (1st && then ||), how to cast signed to unsigned, how to write a readable code, how to avoid compile warnings, how to code at all. This is the first attempt to to fix their mess, affecting some old Ascent code, too.

+++ src/arcemu-world/Creature.cpp
@@ -1202,7 +1202,7 @@ bool Creature::isattackable(CreatureSpawn *spawn){
 }
 
 uint8 get_byte(uint32 buffer, uint32 index){
-	uint32 mask = ~0ul;
+	uint32 mask = uint32(~0ul);
 	if(index > sizeof(uint32)-1)
 		return 0;
